Changement types Long en int pour identifiants des objets et suppression des points d'intérêts

+++ app/src/main/java/fr/iutrodez/a4awalk/modeles/Person.java
@@ -11,5 +11,5 @@ public interface Person {
     Level getNiveau(); // Sportif, Entrainé, Débutant
     Morphology getMorphologie(); // Légère, Moyenne, Forte
 
-    Long getId();
+    int getId();
 }
